refactor: combine the server:user and account modules
Combine the two modules into a datasource:user module that provides all the logic needed for
getting/storing a user data. The next step is to find a way to use the database from within this
module, so it can self contain all of its logic.
