+++ app/src/main/java/com/chesire/nekome/injection/UserModule.kt
@@ -1,11 +1,11 @@
 package com.chesire.nekome.injection
 
+import com.chesire.nekome.datasource.user.remote.UserApi
 import com.chesire.nekome.kitsu.KITSU_URL
 import com.chesire.nekome.kitsu.adapters.ImageModelAdapter
 import com.chesire.nekome.kitsu.user.KitsuUser
 import com.chesire.nekome.kitsu.user.KitsuUserService
 import com.chesire.nekome.kitsu.user.adapter.RatingSystemAdapter
-import com.chesire.nekome.user.api.UserApi
 import com.squareup.moshi.Moshi
 import dagger.Binds
 import dagger.Module
