+++ app/src/test/java/com/chesire/nekome/services/RefreshAuthWorkerTests.kt
@@ -3,9 +3,9 @@ package com.chesire.nekome.services
 import android.content.Context
 import androidx.work.ListenableWorker
 import androidx.work.WorkerParameters
-import com.chesire.nekome.account.UserRepository
 import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.user.UserRepository
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
@@ -15,6 +15,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Test
 
 class RefreshAuthWorkerTests {
+
     @Test
     fun `doWork null userId returns Result#success early`() = runBlocking {
         val mockContext = mockk<Context>()
