+++ app/src/test/java/com/chesire/nekome/services/RefreshSeriesWorkerTests.kt
@@ -3,8 +3,8 @@ package com.chesire.nekome.services
 import android.content.Context
 import androidx.work.ListenableWorker
 import androidx.work.WorkerParameters
-import com.chesire.nekome.account.UserRepository
 import com.chesire.nekome.core.Resource
+import com.chesire.nekome.datasource.user.UserRepository
 import com.chesire.nekome.library.SeriesRepository
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -15,6 +15,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Test
 
 class RefreshSeriesWorkerTests {
+
     @Test
     fun `doWork null userId returns Result#success early`() = runBlocking {
         val mockContext = mockk<Context>()
