+++ features/login/src/main/java/com/chesire/nekome/app/login/syncing/SyncingViewModel.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
-import com.chesire.nekome.account.User
-import com.chesire.nekome.account.UserRepository
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.extensions.postError
 import com.chesire.nekome.core.extensions.postSuccess
 import com.chesire.nekome.core.flags.AsyncState
+import com.chesire.nekome.datasource.user.User
+import com.chesire.nekome.datasource.user.UserRepository
 import com.chesire.nekome.library.SeriesRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.map
