+++ features/login/src/test/java/com/chesire/nekome/app/login/syncing/SyncingViewModelTests.kt
@@ -2,9 +2,9 @@ package com.chesire.nekome.app.login.syncing
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.chesire.nekome.account.UserRepository
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.flags.AsyncState
+import com.chesire.nekome.datasource.user.UserRepository
 import com.chesire.nekome.library.SeriesRepository
 import com.chesire.nekome.testing.CoroutinesMainDispatcherRule
 import io.mockk.Runs
@@ -20,6 +20,7 @@ import org.junit.Test
 
 @ExperimentalCoroutinesApi
 class SyncingViewModelTests {
+
     @get:Rule
     val taskExecutorRule = InstantTaskExecutorRule()
