+++ features/profile/build.gradle
@@ -30,8 +30,8 @@ android {
 }
 
 dependencies {
-    implementation project(":libraries:account")
     implementation project(":libraries:core")
+    implementation project(":libraries:datasource:user")
     implementation project(":libraries:library")
 
     implementation "androidx.appcompat:appcompat:$appcompat_version"
