+++ features/profile/src/main/java/com/chesire/nekome/app/profile/ProfileViewModel.kt
@@ -3,9 +3,9 @@ package com.chesire.nekome.app.profile
 import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
-import com.chesire.nekome.account.UserRepository
 import com.chesire.nekome.core.flags.SeriesType
 import com.chesire.nekome.core.flags.UserSeriesStatus
+import com.chesire.nekome.datasource.user.UserRepository
 import com.chesire.nekome.library.SeriesDomain
 import com.chesire.nekome.library.SeriesRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
