--- libraries/account/build.gradle
@@ -1,49 +1,0 @@
-plugins {
-    id "com.android.library"
-    id "kotlin-android"
-    id "kotlin-kapt"
-}
-
-android {
-    compileSdkVersion sdk_version
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion sdk_version
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro"
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-    implementation project(":libraries:core")
-    implementation project(":libraries:database")
-    implementation project(":libraries:server:user")
-
-    implementation "androidx.appcompat:appcompat:$appcompat_version"
-    implementation "com.chesire.lintrules:lint-gradle:$lintrules_version"
-    implementation "com.chesire.lintrules:lint-xml:$lintrules_version"
-    implementation "com.google.dagger:hilt-android:$hilt_version"
-    implementation "com.jakewharton.timber:timber:$timber_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
-
-    testImplementation project(":testing")
-    testImplementation "io.mockk:mockk:$mockk_version"
-    testImplementation "junit:junit:$junit_version"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
-}
