+++ libraries/kitsu/user/build.gradle
@@ -31,8 +31,8 @@ android {
 
 dependencies {
     implementation project(":libraries:core")
+    implementation project(':libraries:datasource:user')
     implementation project(':libraries:kitsu')
-    implementation project(':libraries:server:user')
 
     implementation "com.chesire.lintrules:lint-gradle:$lintrules_version"
     implementation "com.chesire.lintrules:lint-xml:$lintrules_version"
