--- libraries/server/user/build.gradle
@@ -1,36 +1,0 @@
-plugins {
-    id "com.android.library"
-    id "kotlin-android"
-}
-
-android {
-    compileSdkVersion sdk_version
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion sdk_version
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro"
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-    implementation project(':libraries:core')
-
-    implementation "com.chesire.lintrules:lint-gradle:$lintrules_version"
-    implementation "com.chesire.lintrules:lint-xml:$lintrules_version"
-}
