+++ settings.gradle
@@ -7,7 +7,6 @@ include ':app',
     ':features:settings',
     ':features:timeline',
 
-    ':libraries:account',
     ':libraries:core',
     ':libraries:database',
     ':libraries:datasource:user',
@@ -23,6 +22,5 @@ include ':app',
     ':libraries:server:library',
     ':libraries:server:search',
     ':libraries:server:trending',
-    ':libraries:server:user',
 
     ':testing'
